feat: note hook事件触发环境变量中增加porjectId和commitId #6675
