+++ src/backend/ci/core/common/common-webhook/api-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/pojo/code/WebhookConstant.kt
@@ -177,3 +177,9 @@ const val PIPELINE_WEBHOOK_ISSUE_ACTION = "PIPELINE_WEBHOOK_ISSUE_ACTION"
 // note事件
 const val PIPELINE_WEBHOOK_NOTE_COMMENT = "PIPELINE_WEBHOOK_NOTE_COMMENT"
 const val PIPELINE_WEBHOOK_NOTE_ID = "PIPELINE_WEBHOOK_NOTE_ID"
+const val PIPELINE_WEBHOOK_NOTE_PROJECT_ID = "PIPELINE_WEBHOOK_PROJECT_ID"
+const val PIPELINE_WEBHOOK_NOTE_NOTEABLE_TYPE = "PIPELINE_WEBHOOK_NOTEABLE_TYPE"
+const val PIPELINE_WEBHOOK_NOTE_AUTHOR_ID = "PIPELINE_WEBHOOK_AUTHOR_ID"
+const val PIPELINE_WEBHOOK_NOTE_CREATED_AT = "PIPELINE_WEBHOOK_NOTE_CREATED_AT"
+const val PIPELINE_WEBHOOK_NOTE_UPDATED_AT = "PIPELINE_WEBHOOK_NOTE_UPDATED_AT"
+const val PIPELINE_WEBHOOK_NOTE_URL = "PIPELINE_WEBHOOK_NOTE_URL"
