+++ src/backend/ci/core/common/common-webhook/biz-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/service/code/handler/tgit/TGitNoteTriggerHandler.kt
@@ -28,21 +28,13 @@
 package com.tencent.devops.common.webhook.service.code.handler.tgit
 
 import com.tencent.devops.common.pipeline.pojo.element.trigger.enums.CodeEventType
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_BASE_REF
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_BEFORE_SHA
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_BEFORE_SHA_SHORT
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_COMMIT_AUTHOR
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_COMMIT_MESSAGE
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_EVENT
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_EVENT_URL
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_HEAD_REF
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_ACTION
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_DESC
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_ID
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_IID
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_PROPOSER
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_TITLE
-import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_MR_URL
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_REF
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_REPO_URL
 import com.tencent.devops.common.pipeline.utils.PIPELINE_GIT_SHA
@@ -57,8 +49,14 @@ import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_ISSUE_OWNER
 import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_ISSUE_STATE
 import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_ISSUE_TITLE
 import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_ISSUE_URL
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_AUTHOR_ID
 import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_COMMENT
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_CREATED_AT
 import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_ID
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_NOTEABLE_TYPE
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_PROJECT_ID
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_UPDATED_AT
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_NOTE_URL
 import com.tencent.devops.common.webhook.pojo.code.WebHookParams
 import com.tencent.devops.common.webhook.pojo.code.git.GitNoteEvent
 import com.tencent.devops.common.webhook.service.code.GitScmService
@@ -108,11 +106,18 @@ class TGitNoteTriggerHandler(
         return event.objectAttributes.note
     }
 
+    @SuppressWarnings("ComplexMethod", "LongMethod")
     override fun retrieveParams(event: GitNoteEvent, projectId: String?, repository: Repository?): Map<String, Any> {
         val startParams = mutableMapOf<String, Any>()
         with(event.objectAttributes) {
             startParams[PIPELINE_WEBHOOK_NOTE_COMMENT] = note
             startParams[PIPELINE_WEBHOOK_NOTE_ID] = id
+            startParams[PIPELINE_WEBHOOK_NOTE_PROJECT_ID] = projectId.toString()
+            startParams[PIPELINE_WEBHOOK_NOTE_NOTEABLE_TYPE] = noteableType
+            startParams[PIPELINE_WEBHOOK_NOTE_AUTHOR_ID] = authorId
+            startParams[PIPELINE_WEBHOOK_NOTE_CREATED_AT] = createdAt
+            startParams[PIPELINE_WEBHOOK_NOTE_UPDATED_AT] = updatedAt
+            startParams[PIPELINE_WEBHOOK_NOTE_URL] = url
         }
         if (projectId != null && repository != null) {
             val (defaultBranch, commitInfo) =
@@ -138,25 +143,21 @@ class TGitNoteTriggerHandler(
             startParams[PIPELINE_GIT_COMMIT_MESSAGE] = message
         }
         event.mergeRequest?.apply {
-            startParams[PIPELINE_GIT_HEAD_REF] = target_branch
-            startParams[PIPELINE_GIT_BASE_REF] = source_branch
-            startParams[PIPELINE_GIT_MR_URL] = url ?: event.objectAttributes.url
-            startParams[PIPELINE_GIT_MR_ID] = id.toString()
-            startParams[PIPELINE_GIT_MR_IID] = iid.toString()
-            startParams[PIPELINE_GIT_MR_TITLE] = title
-            startParams[PIPELINE_GIT_MR_DESC] = description ?: ""
+            val mrRequestId = if (repository is CodeGitlabRepository) {
+                iid
+            } else {
+                id
+            }
+            startParams.putAll(
+                WebhookUtils.mrStartParam(
+                    gitScmService = gitScmService,
+                    mrRequestId = mrRequestId,
+                    projectId = projectId,
+                    repository = repository,
+                    homepage = event.repository.homepage
+                )
+            )
             startParams[PIPELINE_GIT_MR_ACTION] = action ?: ""
-
-            val mrInfo = gitScmService.getMergeRequestInfo(
-                projectId = projectId ?: return@apply,
-                mrId = if (repository is CodeGitlabRepository) {
-                    iid
-                } else {
-                    id
-                },
-                repo = repository ?: return@apply
-            ) ?: return@apply
-            startParams[PIPELINE_GIT_MR_PROPOSER] = mrInfo.author.username
         }
         event.issue?.apply {
             startParams[PIPELINE_WEBHOOK_ISSUE_TITLE] = title
